+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.42.0] - 2025-08-22
+
 ### Added
 
 - `tilebox-storage`: Added `USGSLandsatStorageClient` to download landsat data from the USGS Landsat S3 bucket.
@@ -21,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   overwriting the existing task.
 - `tilebox-workflows`: Fixed a bug where the `deserialize_task` function would fail to deserialize nested dataclasses or
   protobuf messages that are wrapped in an `Optional` or `Annotated` type hint.
+- `tilebox-workflows`: Calling `configure_otel_tracing` and `configure_otel_logging` multiple times correctly configures
+  multiple exporters instead of overwriting the existing ones.
 
 ## [0.41.0] - 2025-08-01
 
@@ -248,7 +252,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.41.0...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.42.0...HEAD
+[0.42.0]: https://github.com/tilebox/tilebox-python/compare/v0.41.0...v0.42.0
 [0.41.0]: https://github.com/tilebox/tilebox-python/compare/v0.40.0...v0.41.0
 [0.40.0]: https://github.com/tilebox/tilebox-python/compare/v0.39.0...v0.40.0
 [0.39.0]: https://github.com/tilebox/tilebox-python/compare/v0.38.0...v0.39.0